Remove the chunkBatch window and update the batch window as requested
