+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/processor/stream/window/BatchWindowProcessor.java
@@ -19,6 +19,8 @@ package org.wso2.siddhi.core.query.processor.stream.window;
 
 import org.wso2.siddhi.annotation.Example;
 import org.wso2.siddhi.annotation.Extension;
+import org.wso2.siddhi.annotation.Parameter;
+import org.wso2.siddhi.annotation.util.DataType;
 import org.wso2.siddhi.core.config.SiddhiAppContext;
 import org.wso2.siddhi.core.event.ComplexEvent;
 import org.wso2.siddhi.core.event.ComplexEventChunk;
@@ -26,6 +28,7 @@ import org.wso2.siddhi.core.event.state.StateEvent;
 import org.wso2.siddhi.core.event.stream.StreamEvent;
 import org.wso2.siddhi.core.event.stream.StreamEventCloner;
 import org.wso2.siddhi.core.event.stream.holder.SnapshotableStreamEventQueue;
+import org.wso2.siddhi.core.executor.ConstantExpressionExecutor;
 import org.wso2.siddhi.core.executor.ExpressionExecutor;
 import org.wso2.siddhi.core.executor.VariableExpressionExecutor;
 import org.wso2.siddhi.core.query.processor.Processor;
@@ -39,22 +42,30 @@ import org.wso2.siddhi.core.util.snapshot.state.SnapshotStateList;
 import org.wso2.siddhi.query.api.exception.SiddhiAppValidationException;
 import org.wso2.siddhi.query.api.expression.Expression;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 /**
  * Implementation of {@link WindowProcessor} which represent a Batch Window that aggregate batch of incoming events
- * together.
+ * together or as chunks.
  */
 @Extension(
         name = "batch",
         namespace = "",
         description = "A window that holds an incoming events batch. When a new set of events arrives, the previously" +
                 " arrived old events will be expired. Batch window can be used to aggregate events that comes " +
-                "in batches.",
+                "in batches. If it has the parameter length specified, then batch window process the batch " +
+                "as several chunks.",
         parameters = {
-                // TODO: 7/7/18 Add parameter batch.size.max
+                @Parameter(name = "window.length",
+                        description = "The length of a chunk",
+                        type = {DataType.INT},
+                        optional = true,
+                        defaultValue = "If length value was not given it assign -1 as length and process " +
+                                "the whole batch as once"
+                ),
         },
         examples = {
                 @Example(
@@ -71,7 +82,10 @@ import java.util.Map;
 )
 public class BatchWindowProcessor extends BatchingWindowProcessor implements FindableProcessor {
 
+    private int length = -1;
+    private int count = 0;
     private SnapshotableStreamEventQueue expiredEventQueue = null;
+    private SnapshotableStreamEventQueue currentEventQueue;
     private boolean outputExpectsExpiredEvents;
     private SiddhiAppContext siddhiAppContext;
     private StreamEvent resetEvent = null;
@@ -81,20 +95,26 @@ public class BatchWindowProcessor extends BatchingWindowProcessor implements Fin
                         boolean outputExpectsExpiredEvents, SiddhiAppContext siddhiAppContext) {
         this.outputExpectsExpiredEvents = outputExpectsExpiredEvents;
         this.siddhiAppContext = siddhiAppContext;
+        currentEventQueue = new SnapshotableStreamEventQueue(streamEventClonerHolder);
         if (outputExpectsExpiredEvents) {
             expiredEventQueue = new SnapshotableStreamEventQueue(streamEventClonerHolder);
         }
-        if (attributeExpressionExecutors.length != 0) {
-            throw new SiddhiAppValidationException(
-                    "Batch window should not have any parameters, but found " + attributeExpressionExecutors.length +
-                            " input attributes");
+        if (attributeExpressionExecutors.length == 1) {
+            length = (Integer) (((ConstantExpressionExecutor) attributeExpressionExecutors[0]).getValue());
+        } else if (attributeExpressionExecutors.length == 0) {
+            length = -1;
+        } else {
+            throw new SiddhiAppValidationException("Length batch window should have at most one parameter (<int> " +
+                    "chunkLength), but found " + attributeExpressionExecutors.length + " input attributes");
         }
     }
 
     @Override
     protected void process(ComplexEventChunk<StreamEvent> streamEventChunk, Processor nextProcessor,
                            StreamEventCloner streamEventCloner) {
-        ComplexEventChunk<StreamEvent> outputStreamEventChunk = new ComplexEventChunk<StreamEvent>(true);
+        List<ComplexEventChunk<StreamEvent>> streamEventChunks = new ArrayList<ComplexEventChunk<StreamEvent>>();
+        ComplexEventChunk<StreamEvent> streamEventChunkStore = new ComplexEventChunk<StreamEvent>(true);
+        ComplexEventChunk<StreamEvent> currentEventChunk = new ComplexEventChunk<StreamEvent>(true);
         synchronized (this) {
             long currentTime = siddhiAppContext.getTimestampGenerator().currentTime();
             if (outputExpectsExpiredEvents) {
@@ -102,12 +122,13 @@ public class BatchWindowProcessor extends BatchingWindowProcessor implements Fin
                     while (expiredEventQueue.hasNext()) {
                         expiredEventQueue.next().setTimestamp(currentTime);
                     }
-                    outputStreamEventChunk.add(expiredEventQueue.getFirst());
+                    currentEventChunk.add(expiredEventQueue.getFirst());
+                    if (resetEvent != null) {
+                        currentEventChunk.add(resetEvent);
+                        resetEvent = null;
+                    }
                 }
                 expiredEventQueue.clear();
-            }
-            if (resetEvent != null) {
-                outputStreamEventChunk.add(resetEvent);
             }
 
             //check whether the streamEventChunk has next event before add into output stream event chunk
@@ -120,12 +141,51 @@ public class BatchWindowProcessor extends BatchingWindowProcessor implements Fin
                         expiredEventQueue.add(clonedStreamEvent);
                     } while (streamEventChunk.hasNext());
                 }
-                resetEvent = streamEventCloner.copyStreamEvent(streamEventChunk.getFirst());
+                streamEventChunkStore.add(streamEventChunk.getFirst());
+            }
+
+            while (streamEventChunkStore.hasNext()) {
+                StreamEvent streamEvent = streamEventChunkStore.next();
+                StreamEvent clonedStreamEvent = streamEventCloner.copyStreamEvent(streamEvent);
+                currentEventQueue.add(clonedStreamEvent);
+                count++;
+
+                if (count == length) {
+                    if (currentEventQueue.getFirst() != null) {
+                        if (resetEvent != null) {
+                            currentEventChunk.add(resetEvent);
+                        }
+                        resetEvent = null;
+                        resetEvent = streamEventCloner.copyStreamEvent(currentEventQueue.getFirst());
+                        resetEvent.setType(ComplexEvent.Type.RESET);
+                        currentEventChunk.add(currentEventQueue.getFirst());
+                    }
+                    count = 0;
+                    currentEventQueue.clear();
+                    if (currentEventChunk.getFirst() != null) {
+                        streamEventChunks.add(currentEventChunk);
+                        currentEventChunk = new ComplexEventChunk<StreamEvent>(true);
+                    }
+                }
+            }
+            if (currentEventQueue.getFirst() != null) {
+                if (resetEvent != null) {
+                    currentEventChunk.add(resetEvent);
+                }
+                resetEvent = null;
+                resetEvent = streamEventCloner.copyStreamEvent(currentEventQueue.getFirst());
                 resetEvent.setType(ComplexEvent.Type.RESET);
-                outputStreamEventChunk.add(streamEventChunk.getFirst());
+                currentEventChunk.add(currentEventQueue.getFirst());
+            }
+            count = 0;
+            currentEventQueue.clear();
+            if (currentEventChunk.getFirst() != null) {
+                streamEventChunks.add(currentEventChunk);
             }
         }
-        nextProcessor.process(outputStreamEventChunk);
+        for (ComplexEventChunk<StreamEvent> processStreamEventChunk : streamEventChunks) {
+            nextProcessor.process(processStreamEventChunk);
+        }
     }
 
     @Override
